requirements.txt
smartapi-python
websocket-client
pandas
numpy
datetime
logzero
pyotp
requests